+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"/Users/santiagorios/Desktop/sr-react/src/screens/SignUp/index.tsx":"1","/Users/santiagorios/Desktop/sr-react/src/components/App/index.tsx":"2","/Users/santiagorios/Desktop/sr-react/src/components/I18n/index.tsx":"3","/Users/santiagorios/Desktop/sr-react/src/components/Input/index.tsx":"4"},{"size":3302,"mtime":1613603497083,"results":"5","hashOfConfig":"6"},{"size":476,"mtime":1613603430730,"results":"7","hashOfConfig":"6"},{"size":1120,"mtime":1613603711813,"results":"8","hashOfConfig":"6"},{"size":994,"mtime":1613602088286,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1xg9b1i",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/santiagorios/Desktop/sr-react/src/screens/SignUp/index.tsx",[],"/Users/santiagorios/Desktop/sr-react/src/components/App/index.tsx",[],"/Users/santiagorios/Desktop/sr-react/src/components/I18n/index.tsx",[],"/Users/santiagorios/Desktop/sr-react/src/components/Input/index.tsx",[]]
