+++ package.json
@@ -11,6 +11,7 @@
     "@types/react": "^16.9.53",
     "@types/react-dom": "^16.9.8",
     "apisauce": "^1.1.2",
+    "classnames": "^2.2.6",
     "history": "^4.10.1",
     "i18next": "^13.1.5",
     "node-sass": "^4.14.1",
@@ -53,6 +54,7 @@
   "devDependencies": {
     "@rescripts/cli": "0.0.14",
     "@testing-library/react-hooks": "^3.7.0",
+    "@types/classnames": "^2.2.11",
     "@types/react-router": "^5.1.11",
     "@types/react-router-dom": "^5.1.7",
     "@types/react-spinkit": "^3.0.6",
