+++ src/components/App/index.tsx
@@ -1,10 +1,20 @@
-import React from 'react';
+import React, { useState } from 'react';
+import i18next from 'i18next';
 
-import 'scss/application.scss';
 import SignUp from 'screens/SignUp/index';
+import I18n from 'components/I18n/index';
+
+import styles from './styles.module.scss';
 
 function App() {
-  return <SignUp />;
+  const [language, setLanguage] = useState(i18next.language);
+
+  return (
+    <div className={`${styles.appContainer} row center middle`}>
+      <I18n language={language} setLanguage={setLanguage} />
+      <SignUp />
+    </div>
+  );
 }
 
 export default App;
