+++ src/components/App/styles.module.scss
@@ -1,0 +1,5 @@
+@import 'scss/application.scss';
+
+.app-container {
+  position: relative;
+}
