+++ src/components/Input/index.tsx
@@ -2,6 +2,8 @@ import React, { InputHTMLAttributes } from 'react';
 import { FieldError } from 'react-hook-form';
 import i18next from 'i18next';
 
+import { I18N_CONFIG } from 'constants/index';
+
 import styles from './styles.module.scss';
 
 interface Props extends InputHTMLAttributes<HTMLInputElement> {
@@ -21,8 +23,8 @@ function InputText({ inputRef, labelText, errorMessage, name, ...rest }: Props) 
       <label className={`m-bottom-1 ${styles.label}`} htmlFor={name}>
         {i18next.t(`${labelText}`)}
       </label>
-      <input className={styles.input} ref={inputRef} {...rest} />
-      {errorMessage && <p className={styles.errorMessage}>{errorMessage}</p>}
+      <input className={styles.input} ref={inputRef} name={name} {...rest} />
+      <p className={styles.error}>{errorMessage ? i18next.t(`${I18N_CONFIG.key}:${errorMessage}`) : ''}</p>
     </div>
   );
 }
