+++ src/screens/SignUp/styles.module.scss
@@ -14,35 +14,6 @@
   width: 300px;
 }
 
-.language-button {
-  border-bottom: 1px solid transparent;
-  color: $black;
-  padding: 5px;
-  position: relative;
-  width: 40px;
-
-  &::before {
-    background-color: $black;
-    bottom: -1px;
-    content: '';
-    display: block;
-    height: 1px;
-    left: 50%;
-    position: absolute;
-    transition: all 0.3s;
-    width: 0;
-  }
-
-  &:hover::before {
-    left: 0;
-    width: 100%;
-  }
-}
-
-.active {
-  border-bottom: 1px solid $black;
-}
-
 .form-image {
   height: 70px;
   width: 212.42px;
